Live
on GoolFutsal.com

+++ view/wp-content/plugins/js_composer/include/classes/editors/popups/class-vc-templates-panel-editor.php
@@ -7,7 +7,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  * Class Vc_Templates_Panel_Editor
  * @since 4.4
  */
-class Vc_Templates_Panel_Editor implements Vc_Render {
+class Vc_Templates_Panel_Editor {
 	/**
 	 * @since 4.4
 	 * @var string
@@ -90,11 +90,7 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		if ( 'my_templates' === $category['category'] ) {
 			$category['output'] = '';
 
-			if ( vc_user_access()
-				->part( 'templates' )
-				->checkStateAny( true, null )
-				->get()
-			) {
+			if ( vc_user_access()->part( 'templates' )->checkStateAny( true, null )->get() ) {
 				$category['output'] .= '
 				<div class="vc_column vc_col-sm-12" data-vc-hide-on-search="true">
 					<div class="vc_element_label">' . esc_html( __( 'Save current layout as a template', 'js_composer' ) ) . '</div>
@@ -192,15 +188,11 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		$preview_template_title = esc_attr( 'Preview template', 'js_composer' );
 		$add_template_title = esc_attr( 'Add template', 'js_composer' );
 		$deleteControlRender = '';
-		if ( vc_user_access()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->get()
-		) {
+		if ( vc_user_access()->part( 'templates' )->checkStateAny( true, null )->get() ) {
 			$delete_template_title = esc_attr( 'Delete template', 'js_composer' );
 			$deleteControlRender = <<<DATA
 				<button type="button" class="vc_general vc_ui-control-button" data-vc-ui-delete="template-title" title="$delete_template_title">
-					<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-trash-dark"></i>
+					<i class="vc-composer-icon vc-c-icon-delete_empty"></i>
 				</button>
 DATA;
 		}
@@ -211,12 +203,12 @@ DATA;
 			<div class="vc_ui-list-bar-item-actions">
 				<button type="button" class="vc_general vc_ui-control-button" title="$add_template_title"
 					 	data-template-handler="">
-					<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-add-dark"></i>
+					<i class="vc-composer-icon vc-c-icon-add"></i>
 				</button>
 				$deleteControlRender
 				<button type="button" class="vc_general vc_ui-control-button" title="$preview_template_title"
 					data-vc-container=".vc_ui-list-bar" data-vc-preview-handler data-vc-target="[data-template_id_hash=$template_id_hash]">
-					<i class="vc_ui-icon-pixel vc_ui-preview-icon"></i>
+					<i class="vc-composer-icon vc-c-icon-arrow_drop_down"></i>
 				</button>
 			</div>
 HTML;
@@ -247,11 +239,11 @@ HTML;
 		<div class="vc_ui-list-bar-item-actions">
 			<button type="button" class="vc_general vc_ui-control-button" title="$add_template_title"
 					data-template-handler="">
-				<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-add-dark"></i>
+				<i class="vc-composer-icon vc-c-icon-add"></i>
 			</button>
 			<button type="button" class="vc_general vc_ui-control-button" title="$preview_template_title"
 				data-vc-container=".vc_ui-list-bar" data-vc-preview-handler data-vc-target="[data-template_id_hash=$template_id_hash]">
-				<i class="vc_ui-icon-pixel vc_ui-preview-icon"></i>
+				<i class="vc-composer-icon vc-c-icon-arrow_drop_down"></i>
 			</button>
 		</div>
 HTML;
@@ -264,14 +256,7 @@ HTML;
 	 * vc_filter: vc_templates_render_frontend_template - called when unknown template received to render in frontend.
 	 */
 	function renderFrontendTemplate() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		add_filter( 'vc_frontend_template_the_content', array(
 			&$this,
@@ -325,19 +310,6 @@ HTML;
 	}
 
 	/**
-	 * @since 4.4
-	 * @deprecated 4.7
-	 *
-	 * vc_filter: vc_templates_render_template - hook to override singe template rendering in panel window
-	 */
-	public function render() {
-		_deprecated_function( '\Vc_Templates_Panel_Editor::render', '4.7 (will be removed in 4.11)', '\Vc_Templates_Panel_Editor::renderUITemplate' );
-		vc_include_template( 'editors/popups/panel_templates.tpl.php', array(
-			'box' => $this,
-		) );
-	}
-
-	/**
 	 * @since 4.7
 	 */
 	public function renderUITemplate() {
@@ -352,14 +324,7 @@ HTML;
 	 * @since 4.4
 	 */
 	public function save() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->checkStateAny( true, null )->validateDie();
 
 		$template_name = vc_post_param( 'template_name' );
 		$template = vc_post_param( 'template' );
@@ -401,14 +366,7 @@ HTML;
 	 * vc_filter: vc_templates_render_backend_template - called when unknown template requested to render in backend
 	 */
 	public function renderBackendTemplate() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		$template_id = vc_post_param( 'template_unique_id' );
 		$template_type = vc_post_param( 'template_type' );
@@ -443,17 +401,10 @@ HTML;
 	 * @since 4.8
 	 */
 	public function renderTemplatePreview() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( array(
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( array(
 				'edit_post',
 				(int) vc_request_param( 'post_id' ),
-			) )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+			) )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		$template_id = vc_request_param( 'template_unique_id' );
 		$template_type = vc_request_param( 'template_type' );
@@ -512,29 +463,28 @@ HTML;
 	 * @since 4.4
 	 */
 	public function delete() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->checkStateAny( true, null )->validateDie();
 
 		$template_id = vc_post_param( 'template_id' );
+		$template_type = vc_post_param( 'template_type' );
 
 		if ( ! isset( $template_id ) || '' === $template_id ) {
 			die( 'Error: Vc_Templates_Panel_Editor::delete:1' );
 		}
+		if ( 'my_templates' === $template_type ) {
 
-		$saved_templates = get_option( $this->option_name );
-		unset( $saved_templates[ $template_id ] );
-		if ( count( $saved_templates ) > 0 ) {
-			update_option( $this->option_name, $saved_templates );
+			$saved_templates = get_option( $this->option_name );
+			unset( $saved_templates[ $template_id ] );
+			if ( count( $saved_templates ) > 0 ) {
+				update_option( $this->option_name, $saved_templates );
+			} else {
+				delete_option( $this->option_name );
+			}
+			wp_send_json_success();
 		} else {
-			delete_option( $this->option_name );
+			do_action( 'vc_templates_delete_templates', $template_id, $template_type );
 		}
-		die();
+		wp_send_json_error();
 	}
 
 	/**
@@ -616,6 +566,9 @@ HTML;
 			);
 			$category_templates = array();
 			foreach ( $default_templates as $template_id => $template_data ) {
+				if ( isset( $template_data['disabled'] ) && $template_data['disabled'] ) {
+					continue;
+				}
 				$category_templates[] = array(
 					'unique_id' => $template_id,
 					'name' => $template_data['name'],
@@ -626,8 +579,10 @@ HTML;
 					'custom_class' => isset( $template_data['custom_class'] ) ? $template_data['custom_class'] : false,
 				);
 			}
-			$arr_category['templates'] = $category_templates;
-			$data[] = $arr_category;
+			if ( ! empty( $category_templates ) ) {
+				$arr_category['templates'] = $category_templates;
+				$data[] = $arr_category;
+			}
 		}
 
 		// To get any other 3rd "Custom template" - do this by hook filter 'vc_get_all_templates'
@@ -865,7 +820,6 @@ HTML;
 	 * @todo move to autoload or else some where.
 	 * @since 4.4.3
 	 *
-	 * @return string
 	 */
 	public function addFrontendTemplatesShortcodesCustomCss() {
 		$output = $shortcodes_custom_css = '';
